added about page

+++ Inforce.web-layer/src/app/home/home.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { UrlModel, UrlForShortModel, ShortenerUrl } from '../../models/url';
+import { UrlModel, UrlForShortModel } from '../../models/url';
 import { UrlService } from '../../services/urlService';
-import { Observable } from 'rxjs';
+import { Observable, concatWith } from 'rxjs';
 import { ActivatedRoute, Route, Router } from '@angular/router';
 import { LocalService } from '../../services/localService';
 import { JwtHelperService } from '@auth0/angular-jwt';
@@ -27,22 +27,34 @@ export class HomeComponent implements OnInit {
   code: string = "";
   errorMessage: string = '';
 
+  userId: string = '';
   isAdmin: boolean = false;
   islogged: boolean = false;
 
   getById(id: string) {
     this.urlService.getById(id).subscribe(data => {
       this.router.navigate(['/info', data]);
-    })
+    }
+      , error => {
+        this.errorMessage = error.message;
+      })
   }
 
   deleteById(id: string): void {
-    this.urlService.deleteById(id).subscribe(data => {
-      this.urls$ = this.urlService.getAll();
-    },
-      error => {
-        this.errorMessage = error.message;
-      });
+
+    this.urlService.getCreatorIdByUrl(id).subscribe(creatorId => {
+      if (creatorId === this.userId || this.isAdmin) {
+        this.urlService.deleteById(id).subscribe(data => {
+          this.urls$ = this.urlService.getAll();
+        },
+          error => {
+            this.errorMessage = error.message;
+          });
+      }
+      else {
+        this.errorMessage = "You can't delete notes that aren't yours";
+      }
+    });
   }
 
   deleteAll() {
@@ -58,6 +70,7 @@ export class HomeComponent implements OnInit {
     let url = new UrlForShortModel();
     url.longUrl = this.longUrl;
     url.code = this.code;
+    url.userId = this.userId;
 
     this.urlService.add(url).subscribe(data => {
       this.urls$ = this.urlService.getAll();
@@ -75,6 +88,7 @@ export class HomeComponent implements OnInit {
       this.islogged = true;
       let decodedData = this.jwtHelperService.decodeToken(token);
       this.isAdmin = decodedData.role == "Admin";
+      this.userId = decodedData.jti;
     }
   }
